refactor: drop domain media ScreenshotRepository interface (#129)
Move screenshot repo boundary to usecase-local screenshotRepo.
ScreenshotFilter stays on domain entities.
